Simplified Package Loading API
This is a re‐working of #2166, but branched from the tip of master.

One static function each has been added to Manifest, Package and
PackageGraph that allows them to be loaded by supplying only the
location of swift and the location of the package to load.  This
necessitated some minor refactoring of Workspace internals, so that code
could be shared instead of duplicated.  The package-info example has
been updated to use the simpler APIs.  For package clients, this amounts
to an ergonomic reduction of about 45 lines down to 1, whenever no
customization is needed.

(cherry picked from commit bf7c24b9e86d7bf14652ebd094129984208739e8)