Add .swiftpm/xcode to default gitignore
If you use the new SwiftPM + Xcode 11 integration and open the scheme
viewer, Xcode writes a shared scheme. This shared scheme ends up in a
path that wasn't previously covered by this gitignore since it's
normally intended to be checked in. We can't ignore the entire .swiftpm
directory since this can contain other files, but this ignores all Xcode
created artifacts.