[Workspace] Improve how we handle changes to Package.resolved
This improves our resolution logic when we already have state and there
are updates to Package.resolved file. Previously, we were restoring the
Package.resolved file with our current state if we saw that there is
a missing entry in there. This is not correct since dependencies can
disappear as part of an update performed on some other machine (or just
switching between commits/branches without removing state).

<rdar://problem/51080676> [SR-10718]