[Commands] Add build support for multiroot packages
This adds a --multiroot-data-file option which will allow building
multiroot packages using swift-build. The code in SwiftPM was always
written to support this but it was never exposed via the command-line.
With the recent Xcode integration, it seems useful to expose this option
for debugging purposes at least. For convenience, this option will read the
xcworkspace file which is used by Xcode.

Note that this is not an official feature of command-line SwiftPM yet so
there will be an unsilencable warning when this option is used.

<rdar://problem/53722817>