Refactor DiagnosticsEngine
This greatly simplifies how diagnostics are created and emitted in
SwiftPM. The old statergy was a little complex and it didn't pull off
its weight. It was written before ExpressibleByStringInterpolation was
redesigned so it didn't really make sense to continue with this model.
There was also a big overhead for creating and emitting new diagnostics
as each diagnostic required a bunch of boilerplate setup code.

This PR redesigns the class with inspiration from SwiftSyntax's
DiagnosticsEngine. Diagnostics are now cheap to create and easy to emit.
In the general case, it isn't necessary to create a dedicated type for
individual diagnostic but there is a facility to do when necessary. We
can also attach an unique ID to individual untyped diagnostic by using
Swift's #function but I'll avoid that until we really need it.

There is still some cleanup work left to do but that will come in later PRs.

<rdar://problem/49466682>