Implement build manifest caching
This implements caching for llbuild manifest behind
--enable-build-manifest-caching flag. This currently only works for
`swift build` invocations but there is enough facility to extend this to
other tools (we need to store more information in BuildDescription).

This works by creating a new command "PackageStructure" which depends on
the directory structure of the package, the package manifest,
Package.resolved file, build parameters and the process env.

<rdar://problem/53901583>

Here are some null build timings with and without build manifest caching (on 6 core/i9):

| Package   | without | with  |
|-----------|---------|-------|
| trivial   |  0.076  | 0.060 |
| swift-nio |  0.198  | 0.102 |
| vapor     |  0.668  | 0.192 |
| unnamed   |  1.177  | 0.216 |