[PubGrub] Handle edited/local/branch dependencies using package overriding facility
This implements properly handling unversioned and branch-based
dependencies using a "package overriding" facility in the resolver. The
main advantage of this approach is that we don't need to backtrack for
branch/unversioned dependencies since it seems way too hard to apply set
relations to these constraints. The fact that unversioned and
branch-based dependencies have an overriding behavior makes using set
relations even harder. We might want to do that at some point but this
approach works well and provides accurate results right now.

<rdar://problem/46075343>