TemporaryFile: simplify clean up
`FileManger.removeItem(atPath:)` will remove the file or directory.
`rmdir` will only remove the directory if it is empty, which is the
default behaviour unless an explicit delegate is provided to the
FileManager.  Collapse the two cases into one which enables this path to
work on Windows which does not provide `rmdir`.