[Pubgrub] Plumb PinStore to the resolver
Since PubGrub is capable of handling the data from Package.resolved
directly, we don't need to send precomputed set of valid pins. This
moves the resolver towards a much more cleaner model and we should be
able to get rid of "valid pins" computation when we remove the old
resolver.