[PubGrub] Handle Package.resolved for version-based dependencies
This implements support for handling pinned versions inside
Package.resolved for version-based dependencies. Unlike the old
resolver, PubGrub can natively handle pinned dependencies and
automatically throw away the pins that are not valid any more due to
change in constraints.

Note that this doesn't implement the pin support for branch-based
dependencies. That'll be coming in the next commit.

<rdar://problem/54100127>