[PubGrub] Handle Package.resolved for branch-based dependencies
This will ensure that we respect the SHA in Package.resolved for
branch-based dependencies during resolution. We don't really need to do
anything for commit-based dependencies since they are literally pinned
in the Package.swift file.

This actually fixes a long standing bug that we've had in the old
resolver where we wouldn't respect the revision recorded in
Package.resolved for branch-based dependencies during resolution time
(47746146).