[PubGrub] [NFC] Create PubGrubPackageContainer class
This class is responsible for handling PubGrub specific logic for
an individual package container. For example, this has the method for
getting best version of a package and computing incompatibilities at
a given version.

This should also allow us to nicely encapuslate the logic for computing
bounds of a package dependency declaration and controlling when external
incompatibilities are emitted.