[PubGrub] Implement bound computation of incompatibilities added during decision making
This will implement bound computation of a package dependency when its
being added as an incompatibility during decision making. The basic idea
is to reduce the amount of incompatibilities that are added so pubgrub
is efficiently able to reason about multiple versions of packages.

Bound computation can be costly since we have to iterate the versions
before and after the chosen version but it is an important part of the
algorithm. We might we might need to figure out a way to counter balance
that cost depending on the impact (which I'll measure separately as part
of performance testing).

<rdar://problem/54099969>