Restructure project for tools support core repository
This renames Basic, Utility and their dependencies to TSCBasic and TSCUtility.
This makes it possible to copy their directories to the tools-support-core repository to be able to use them in other projects too.
TestSupport is now divided in TSCTestSupport and SPMTestSupport where the latter contains all SwiftPM specific parts and TSC the generic parts which use the tools support core interfaces.

rdar://55022772