Remove @testable from codebase
Remove the usage of @testable imports because they force us to
enable testability in release mode which turns off compiler optimization
that we could be getting without testability. Instead, the APIs should
be public in the first place or we can prefix them with _ to indicate
that they're not expected to be used by the clients. This also allows us
to do proper performance testing.

<rdar://problem/58052876> Remove the usage of @testable from SwiftPM