Properly conditionalize dependencies in Package.swift
Some of the logic for conditionalization was wrong in our
manifest/bootstrap script. For example, tools support core is cloned
when SWIFTCI_USE_LOCAL_DEPS is set but it's not set when linking llbuild
as a framework. Similarly, we should turn SWIFTPM_BOOTSTRAP to something
better that describes its purpose.

<rdar://problem/58303769>