Start sketching out integration with Xcode's build system
The resources proposal (SE-0271) will allow Swift packages to contain various
kinds of file types that were previously unsupported. This patch starts work in
a direction where command-line swiftpm will be able to directly communicate
with the Xcode build system. This will allow swiftpm command-line tools to work
with packages that contain Apple specific resource files. This is important for
usability of command-line tools for macOS users. We could also use the
generated project support and then delegate to xcodebuild but a direct
communication with the build system will provide a deeper integration
experience.

Xcode's build system has a command-line tool that can be used to perform
builds. This is not officially documented but the build system team at Apple
will support command-line swiftpm as a client. The build system operates on
Project Interchange Format (PIF), which is a structured representation of the
project model. The xcbuild tool can be invoked with the PIF and some
command-line operations to perform the build.

This patch can generate a trivial PIF for a package with an executable
product. An example invocation for an executable package "foo" would be:

$ alias xcbuild=$(xcode-select -p)/../SharedFrameworks/XCBuild.framework/Versions/A/Support/xcbuild
$ swift package dump-pif > foo.pif
$ xcbuild build foo.pif --target foo --configuration Debug --derivedDataPath
$PWD/.build

<rdar://problem/58646702>