[SPMLLBuild] Gracefully handle missing data
This will throw instead of crashing if we can't form a build value.
That can happen if we add new properties to the objects that get encoded in
the database but not bump the schema. This change will cause the build
value to be out-of-date and instead force recomputation of the cached
value.

<rdar://problem/51146826>