[PackageLoading] Handle header files in TargetSourcesBuilder
Since header files are allowed to be mixed with sources, it makes sense
to handle them inside the target sources builder. This allow us to not
special case the public headers directory which can lead to weird edge
cases with previously supported layouts. This patch also adds two more
things 1) diagnostics for duplicate declaration in sources parameter 2)
move the explicit resources check to the vNext tools version as
resources isn't a Swift 5.2 feature.

<rdar://problem/56783465>