Add a user-friendly diagnostic for compute-checksum command. (#2544)
* Add a user-friendly diagnostic for compute-checksum command.

* Move to using Archiver's supported extensions.

* Re-introduce the file check, further tests.

This also expands the tests to include a supported extension that
doesn't exist, as well as a supported extension that's actually a
directory.

* Changed file names to be more accurate.

* Fixed PackageToolTest test.

* Properly fixes test.

Previously it was testing the checksum of a `.md` file, but that won't
work anymore with the improvements. This adds a `.zip` file of the
`.xcframework` and tests it. The `zip` file was created with `zip -r
MyFwk.zip MyFwk.xcframework`.