Make sure main and test targets are always created
If a package contains only test targets (like Integration Tests), SwiftPM currently fails building it because it generates a manifest with a `defaultTarget` of `main`, but no target called `main` exists as they are lazily created. This change makes it so that SwiftPM always creates `main` and `test` targets, even if they end up empty.