[PackageGraph] Fix package name validation for product target dependencies
The package name validation was wrong for product target dependencies
(`.product(name: "<product>", package: "<package>")`) as we were looking
up the package using its identity instead of the name. This completely
breaks package loading in 5.2 if a package wants to use a product from
a dependency that doesn't match the package name of that dependency.

<rdar://problem/59744465>