[PackageModel] Add sdk options to platform description
This adds a way to express generic "sdk options" in string version of
the platforms API. The idea is that a package might be able to express
some platform-specific options for the underlying build system in
a generic way. Currently, this is just for experimentation purposes and
we might want some other (better, more typesafe) way of allowing
extending the package description language.

<rdar://problem/59788403>