Fix InMemoryGitRepository by expecting its contents to present at root
`InMemoryGitRepository` is initialized with a path and a in-memory file system where it belongs. It then creates seperate in-memory file systems for each of its commits.

Previously, the contents of the repository in the commit file systems was at the same location as in the working directory file system, which caused issues when cloning the repository to a different location on the file system (in installHead).

Now, the contents of the repository in the commit file systems is located at root.