[ManifestLoader] Capture JSON output from a manifest on a separate pipe than stdout
Manifests can emit output to stdout (useful for debugging purposes).  When switching from an interpreted to a compiled manifest, this output was no longer captured; and more seriously, if enough stdout output was emitted to fill the stdout buffer (or if buffering was disabled), the stdout output would interfere with the emitted JSON.

This change causes the JSON to again be emitted on a separate channel from the stdout, as it was when the manifest was interpreted.