Amending to use a file and a different command line argument, based on PR feedback.  This argument to the compiled manifest is not API of any kind (not even private API), so there is no issue with changing it.
There are possible various error conditions that the existing code doesn't handle, and this change doesn't improve on that (it also doesn't make it worse).  A future change should consider making those cases more robust, although it may be of limited practical use, since this is an internal contract between two pieces of code.  However, it would be good to have better diagnostics in the case of I/O errors or other environmental problems.