[SR-12553] [ManifestLoader] Capture JSON output from a manifest separately from the stdout (#2683)
Manifests can emit output to stdout (useful for debugging purposes).  The earlier switch from an interpreted to a compiled manifest caused this output to no longer be captured; and more seriously, if enough stdout output was emitted to fill the stdout buffer (or if buffering was disabled), then the stdout output would interfere with the emitted JSON description of the package manifest.

This change causes the JSON to again be emitted to a separate file so that there is no interference with stdout output, restoring the behavior of when the manifest was interpreted.

There are possible various error conditions that the existing code doesn't handle, and this change doesn't improve on that (it also doesn't make it worse).  A future change should consider making those cases more robust, although it may be of limited practical use, since this is an internal contract between two pieces of code.  However, it would be good to have better diagnostics in the case of I/O errors or other environmental problems.