[SR-12613] [ManifestLoader] Go back to using `-fileno` option for JSON output, to avoid revlock between libSwiftPM and PackageDescription
This restores the functionality prior to SR-12553, while also keeping the separation between stdout and the JSON output emitted by the package manifest.  It still uses a file in the file system for the JSON, but goes back to having it be opened by the caller of the manifest, passing the file descriptor down to the manifest process as before the switch from interpreted to compiled manifests.

Unit test added for SR-12553 still provides test coverage.