Add support for the Scudo sanitizer. (#2717)
* Add support for the Scudo sanitizer. + Test sanitizers with build plans

The Swift PR apple/swift#28538 adds support for the Scudo hardened
allocator to swiftc. This patch adds support for it to Swift Package
Manager, allowing the Package Manager to correctly plumb Scudo through
the entire build pipeline.

See the aforementioned PR for more discussion on the value of supporting
Scudo.

(cherry picked from commit 594ccaa049d35ffaf09ccab6ce596d0a5a546a25)

* Test sanitizers with build plans. (#2720)

The current tests for sanitizers in SwiftPM are essentially "live": they
build a package that will do something to trigger a sanitizer, and then
run it and confirm it fails. This is a pretty fragile way to test, and
it exposes SwiftPM to a number of possible test failures that don't
have anything to do with the core mission (exposing the sanitizer).

This patch replaces those tests with build plan tests that validate that
SwiftPM correctly invokes swiftc and clang. These tests are acceptable
because for now SwiftPM doesn't have any smarts with the sanitizers: it
just passes them all through to swiftc and clang.

This patch also deletes the current live tests, as they are no longer
necessary.

(cherry picked from commit 984618b84a533da40c9aff6964329f10288ef150)