[Build] Add support for content-based file tracking
This adds experimental support for content-based file tracking behind
`--experimental-track-content-hash` flag. This is currently implemented
by mangling the file hash into some of the stat fields used by llbuild
but we can move to the proper API once it's available. This feels a bit
awkward but it allows us to immediately start experimenting this
approach on Swift packages.