[PackageLoading] Add diagnostics for missing resources
An improvement was suggested under SR-13084 in which we should display a diagnostics warning to the user should they try to add a resource with a path which doesn't resolve.

I've chosen to make this a warning instead of an error in order to not be a breaking change. Though both could be considered valid.