Some of the GenerateXcodeprojTests are hanging on pull request testing
Some of the Xcode project generation tests were using in-memory file systems and others not.  But the Xcode project generation code does not currently support non-local file systems (which should itself perhaps be addressed), and so this could cause tests to start traversing large parts of the actual file system, including over NFS mounts.

Ideally the Xcode project generation should support the in-memory file system, but since this functionality was primarily useful before Xcode supported loading packages natively, it's not clear whether it's worth the effort to fix that.

rdar://65200866