[Build] Generate unavailable accessor for Bundle.module when no bundle is found
An improvement was suggested under SR-13084 in which we should display a message to the user should they try to access Bundle.module without correctly configuring SPM resources.

Previously if no resource bundle was created (and therefore no `bundlePath`)  we would skip over the code generation for this module.

With this change we still create the static variable but set it to unavailable through the declaration which will alert the user of the misuse hopefully pointing them in the right direction.