Remove warning and update the documentation to reflect the long-standing behavior of treating all C module headers as umbrella headers by default
After more discussion about this, we will be removing the warning and just updating the documentation to match the behavior that’s been in place since 2016.  This is because the current implemented behavior was, according to the extensive PR discussion, the intended behavior, and there is concern about how many existing packages would require changes to get rid of the warning.

We will put up an evolution proposal with refined module map semantics, and at that time, probably tie it to the tools version, since this has now been the semantic for four years and a lot of packages have been written in the meantime.

rdar://65692136