[5.3] Update the documentation to reflect the long-standing behavior of treating all C module headers as umbrella headers by default
The currently documented behavior for how the headers in the  directory are handled doesn't match what actually happens.  Since the current behavior dates back to the original implementation and many packages rely on it, we keep the existing semantics but correct the documentation.  We may want to adjust the current semantics in a future evolution proposal.

rdar://problem/65751265