[SR-13235] libSwiftPM should vend a ClangTarget's module map generation info
This also modifies module map generation to use the unified logic for determining the module map type, and cleans up diagnostics reporting.

With detection of module map type (based on a target's header layout) separate from generation of a module map, the existing ModuleMapGenerator API of having a struct that gets initialized only to perform the one operation (either determining module map type of generating a module map) is a bit clunky.  We may want to refine that to make these standalone functions.

Vending the target's module map layout information provides more semantic information to the client than just vending the raw contents of the generated module map.  This semantic information can then (for example) be shown in the user interface of an IDE, etc.

rdar://65678318