[SR-13334] Targets should record the tools version of the package that declared them
A package's declared swift tools version determines not only the API that is available to the package manifest, but also the semantics used for the target at build time.  This can include how a Clang target's headers are represented in an automatically generated module map, etc.

Currently, targets in the SwiftPM package model don't record the declared tools version of the package manifest from where they came.  This plumbs that information through from the package, so that it is available to the build plan and libSwiftPM clients.

There is no semantic change in this PR, but it will enable fixes where the semantics depend on the declared tools version.