Reinstate manifest sandboxing (#2848)
* Reinstate manifest sandboxing

Manifest loading has been sandboxed on macOS for a while, but the change
in #2518 broke it for 5.3.

https://bugs.swift.org/browse/SR-13346
rdar://problem/66586184

* Test manifest sandboxing

This adds tests for manifest sandboxing. Since we made the sandbox more
restrictive in 5.3, there are two tests:
- in 5.2 we check if we can make a network request
- in 5.3 we check if we can launch a process

* Fix 5.3 sandbox profile

* Use 127.0.0.1 for the attempted network request

It doesn't actually matter whether or not the request would succeed
since the test assumes it will always be blocked.