Workspace: assume that the user is always correct
When the user specifies the Swift executable via `SWIFT_EXEC` in the
environment, assume that the value is supposed to be used.  This changes
the order in which the Swift compiler is evaluated as being present.
This can avoid an unnecessary lookup for the executable in the case that
it is specified by the user.