PackageLoading: escape `\` when generating the modulemap
The `pathString` is overloaded and used as both a printing string and a
path string.  When used to generate the module map, we need to escape
the path separator as it is an escape sequence indicator otherwise.

This allows us to generate modulemaps which can be consumed by clang
on Windows.