SPMBuildCore: detect Windows hosts properly
This ensures that SPM identifies the host correctly enabling platform
specific linker and compile settings to be applied properly.
Previously, we would treat Windows as Linux resulting in incorrect
application of linker options.  This was detected when building
swift-package-manager with swift-package-manager on Windows and we would
attempt to link against `libm`.