PackageLoading: introduce `-Xmanifest`
This introduces a new `-Xmanifest` flag that will allow the user to pass
additional flags to the manifest build invocation.  This is particularly
important for the case of Windows, where the split SDK/toolchain model
requires the additional flags such as `-sdk` or `-resource-dir` to be
passed to the manifest build system.