[SR-13484] Use `--no-checkout` when cloning packages, since it's immediately followed by a `checkout` anyway
SwiftPM fetches a remote repository by mirroring it to the local file system, then clones that local repository to a working directory, after which it checks out a specific ref.

When the local clone happens, Git also does a check out.  This is unnecessary, since the checkout will immediately be replaced by the intended ref, and it also makes the checkout sensitive to whether the default branch name is set up correctly for the repository.

Since the working copy is guaranteed to be checked out after the clone, there is no need to do a checkout as part of the cloning.

rdar://68165300