PackageLoading,Workspace: derive flags for Windows
Use the special environment variable named `SDKROOT` which provides the
root of the SDK as a means for deriving the flags for building code on
Windows.  We assume that the layout is in the following canonical
layout (subject to incompatible evolution):

```
  Developer                                         (%DEVELOPER_DIR%)
    `- Platforms
        `- Windows.platform
            +- Developer
            |   +- Library
            |   |    `- XCTest-[Version]
            |   `- SDKs
            |       `- Windows.sdk                  (%SDKROOT%)
            |           +- SDKSettings.plist
            `- Info.plist
```

This adds 3 sets of flags to the `Xmanifest` options by default:
- `-sdk` which should theoretically be sufficient for everything
- `-I`, `-L`, `-resource-dir` which serve as workarounds for bugs in the
  SDK and swift driver
- `-libc MD` by inspecting the contents of SDKSettings.plist which
  contains the `DEFAULT_USE_RUNTIME` entry in the DefaultSettings tree
  of the plist.

It also adds 4 sets of flags to the default UserToolchain on Windows:
- `-sdk` which should theoretically be sufficient for the resource dir,
  include paths, and library search paths.
- `-I`, `-L`, `-resource-dir` which serve as workarounds for bugs in the
  SDK and swift driver
- `-libc MD` by inspecting the contents of SDKSettings.plist which
  contains the `DEFAULT_USE_RUNTIME` entry in the DefaultSettings tree
  of the plist.
- An additional library and include path by inspecting the Info.plist
  which contains a XCTEST_VERSION entry in the DefaultSettings tree of
  the plist.

Because the platform Info.plist is expected to be provided by the
packaging and is generated at build time, we expect it to be in sync
with the contents of the platform bundle.  The SDKSettings.plist
similarly is expected to be in sync with the SDK that it corresponds to.
By using these auxiliary files we can derive the correct flags for the
SDK that the user is using.

The expectation is that the user is using the same runtime as the SDK
for the toolchain when building by default.  If this does not hold, then
the user must use a `-destination` with the appropriate configuration to
ensure that the build is correct.  However, the build of the manifest
would still use the `SDKROOT` environment variable to determine the
correct configuration of the runtime.